use getDataURIsFromImgURLs
